refactor: Extract DB clearing method
